+++ lib/api/HttpServer.ts
@@ -10,7 +10,7 @@ import * as pjson from "../../package.json";
 import Zamza from "../Zamza";
 import { HttpConfig } from "../interfaces";
 import { routeRoot, routeTopicConfig, routeInfo,
-         routeFetch, routeProduce } from "./routes";
+         routeFetch, routeProduce, routeManage } from "./routes";
 import AccessControll from "./AccessControll";
 
 const DEFAULT_PORT = 1912;
@@ -55,6 +55,7 @@ export default class HttpServer {
         app.use("/api/info", routeInfo(this.zamza));
         app.use("/api/fetch", routeFetch(this.zamza));
         app.use("/api/produce", routeProduce(this.zamza));
+        app.use("/api/manage", routeManage(this.zamza));
 
         this.server = await new Promise((resolve, reject) => {
             let server: any = null;