+++ lib/api/routes/index.ts
@@ -3,3 +3,4 @@ export * from "./topic-config";
 export * from "./info";
 export * from "./fetch";
 export * from "./produce";
+export * from "./manage";