+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # zamza CHANGELOG
 
+## 2018-11-16, Version 0.6.0
+
+* handle message now respects tombstones on compact topics
+* /api/manage endpoints to create and delete keyindexes
+* produce endpoint now respects topic access tokens
+
 ## 2018-11-15, Version 0.5.0
 
 * convenience root endpoints